+++ apps/katalog/src/views/BistaDistributionView.vue
@@ -56,6 +56,7 @@ const schueler = ref([
     title="8a Deutsch"
     :mittelwert="395"
     @schueler-gewaehlt="s => console.log(s)"
+    data-thema="hell"
   />
 </template>`,
   apiEndpoints: [
@@ -229,6 +230,7 @@ const tooltipDemos = computed(() => [
             :mittelwert="mittelwert"
             @schueler-gewaehlt="(s) => (gewaehlt = s)"
             @zone-gewaehlt="() => (gewaehlt = null)"
+            data-thema="hell"
           />
         </div>
         <p v-if="gewaehlt" class="gewaehlt-hinweis" data-testid="schueler-gewaehlt">